requirements.txt
aiohttp==3.9.1
websockets==12.0
python-dotenv==1.0.0
asyncio-mqtt==0.16.1
pandas==2.1.4